scripts/requirements.txt
pillow
numpy
scikit-learn
tensorflow==2.16.1
faiss-cpu
